+++ pkg/app/handler.go
@@ -1,4 +1,4 @@
-package http
+package app
 
 import (
 	"bitcoinrateapp/pkg/core"
